// packages/destination-actions/src/destinations/snowflake/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Settings {
  /**
   * Your Snowflake Account
   */
  account: string
  /**
   * Your Snowflake Username
   */
  username: string
  /**
   * Your Snowflake Password. This is a secret field.
   */
  password: string
  /**
   * Your Snowflake Application Name
   */
  application: string
}